requirements.txt
Flask
SQLAlchemy
Flask-Login
Flask-WTF
email-validator
Pillow>=9.0.0
Flask-APScheduler
